+++ client/reducers/userReducer.js
@@ -4,8 +4,8 @@ import initialState from './initialState';
 
 export default function itemReducer(state = initialState.user, action) {
   switch (action.type) {
-    case actionTypes.LOGIN_SUCCESS: {
-      return { ...action.user.user, token: action.user.token, loggedIn: true };
+    case actionTypes.GET_LOGGED_IN_USER_SUCCESS: {
+      return { ...action.data.user, token: action.data.token, loggedIn: true };
     }
     case actionTypes.LOGOUT_SUCCESS: {
       return {};